chore: rename package http
