+++ gravitee-inference-rest/src/main/java/io/gravitee/inference/rest/http/CustomHttpConfig.java
@@ -1,7 +1,6 @@
-package io.gravitee.inference.rest.customHttp;
+package io.gravitee.inference.rest.http;
 
 import io.gravitee.inference.rest.RestConfig;
-import io.gravitee.inference.rest.openai.OpenaiConfig;
 import io.vertx.core.http.HttpMethod;
 import java.net.URI;
 import java.util.Map;
@@ -11,20 +10,12 @@ public class CustomHttpConfig extends RestConfig {
 
   private final HttpMethod method;
   private final Map<String, String> headers;
-  private final String contentType;
   private final String requestBodyTemplate;
 
-  public CustomHttpConfig(
-    URI uri,
-    HttpMethod method,
-    Map<String, String> headers,
-    String contentType,
-    String requestBodyTemplate
-  ) {
+  public CustomHttpConfig(URI uri, HttpMethod method, Map<String, String> headers, String requestBodyTemplate) {
     super(uri);
     this.method = method;
     this.headers = headers;
-    this.contentType = contentType;
 
     this.requestBodyTemplate = requestBodyTemplate;
   }
@@ -35,10 +26,6 @@ public class CustomHttpConfig extends RestConfig {
 
   public Map<String, String> getHeaders() {
     return headers;
-  }
-
-  public String getContentType() {
-    return contentType;
   }
 
   public String getRequestBodyTemplate() {
@@ -53,13 +40,12 @@ public class CustomHttpConfig extends RestConfig {
     return (
       Objects.equals(method, that.method) &&
       Objects.equals(headers, that.headers) &&
-      Objects.equals(contentType, that.contentType) &&
       Objects.equals(requestBodyTemplate, that.requestBodyTemplate)
     );
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(super.hashCode(), method, headers, contentType, requestBodyTemplate);
+    return Objects.hash(super.hashCode(), method, headers, requestBodyTemplate);
   }
 }
