+++ gravitee-inference-rest/src/main/java/io/gravitee/inference/rest/http/GraviteeInferenceCustomHttpException.java
@@ -1,4 +1,4 @@
-package io.gravitee.inference.rest.customHttp;
+package io.gravitee.inference.rest.http;
 
 public class GraviteeInferenceCustomHttpException extends RuntimeException {
